Re-organize components to add customization controls to more blocks
