+++ assets/src/blocks/amp-story-cta/edit.js
@@ -12,43 +12,19 @@ import {
 	Component,
 	Fragment,
 } from '@wordpress/element';
-import { compose } from '@wordpress/compose';
 import {
 	Dashicon,
 	IconButton,
-	PanelBody,
-	withFallbackStyles,
 } from '@wordpress/components';
 import {
 	URLInput,
 	RichText,
-	ContrastChecker,
-	InspectorControls,
-	withColors,
-	PanelColorSettings,
-	withFontSizes,
-	FontSizePicker,
 } from '@wordpress/editor';
 
 /**
  * Internal dependencies
  */
-import { FontFamilyPicker } from '../../components';
-import { maybeEnqueueFontStyle } from '../../helpers';
 import './edit.css';
-
-const { getComputedStyle } = window;
-
-const applyFallbackStyles = withFallbackStyles( ( node, ownProps ) => {
-	const { textColor, backgroundColor } = ownProps.attributes;
-	const editableNode = node.querySelector( '[contenteditable="true"]' );
-	const computedStyles = editableNode ? getComputedStyle( editableNode ) : null;
-
-	return {
-		fallbackBackgroundColor: backgroundColor || ! computedStyles ? undefined : computedStyles.backgroundColor,
-		fallbackTextColor: textColor || ! computedStyles ? undefined : computedStyles.color,
-	};
-} );
 
 class CallToActionEdit extends Component {
 	constructor( props ) {
@@ -74,68 +50,20 @@ class CallToActionEdit extends Component {
 			attributes,
 			backgroundColor,
 			textColor,
-			setBackgroundColor,
-			setTextColor,
-			fallbackBackgroundColor,
-			fallbackTextColor,
 			setAttributes,
 			isSelected,
 			className,
 			fontSize,
-			setFontSize,
 		} = this.props;
 
 		const {
 			text,
 			url,
-			ampFontFamily,
 		} = attributes;
 
 		return (
 			<Fragment>
 				<div className={ className } ref={ this.bindRef }>
-					<InspectorControls>
-						<PanelBody title={ __( 'Text Settings', 'amp' ) }>
-							<FontFamilyPicker
-								name={ ampFontFamily }
-								onChange={ ( value ) => {
-									maybeEnqueueFontStyle( value );
-									setAttributes( { ampFontFamily: value } );
-								} }
-							/>
-							<FontSizePicker
-								value={ fontSize.size }
-								onChange={ setFontSize }
-							/>
-						</PanelBody>
-						<PanelColorSettings
-							title={ __( 'Color Settings', 'amp' ) }
-							colorSettings={ [
-								{
-									value: backgroundColor.color,
-									onChange: setBackgroundColor,
-									label: __( 'Background Color', 'amp' ),
-								},
-								{
-									value: textColor.color,
-									onChange: setTextColor,
-									label: __( 'Text Color', 'amp' ),
-								},
-							] }
-						>
-							<ContrastChecker
-								{ ...{
-									// Text is considered large if font size is greater or equal to 18pt or 24px,
-									// currently that's not the case for button.
-									isLargeText: false,
-									textColor: textColor.color,
-									backgroundColor: backgroundColor.color,
-									fallbackBackgroundColor,
-									fallbackTextColor,
-								} }
-							/>
-						</PanelColorSettings>
-					</InspectorControls>
 					<RichText
 						placeholder={ __( 'Add text…', 'amp' ) }
 						value={ text }
@@ -174,8 +102,4 @@ class CallToActionEdit extends Component {
 	}
 }
 
-export default compose( [
-	withColors( 'backgroundColor', { textColor: 'color' } ),
-	withFontSizes( 'fontSize' ),
-	applyFallbackStyles,
-] )( CallToActionEdit );
+export default CallToActionEdit;
