+++ assets/src/blocks/amp-story-post-date/index.js
@@ -7,6 +7,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { getMetaBlockSettings } from '../../helpers';
+import './edit.css';
 
 export const name = 'amp/amp-story-post-date';
 
@@ -20,7 +21,7 @@ export const settings = {
 		__( 'date', 'amp' ),
 	],
 	...getMetaBlockSettings( {
-		tagName: 'span',
+		tagName: 'div',
 		attribute: 'date',
 		isEditable: false,
 	} ),
