+++ assets/src/blocks/amp-story-post-title/edit.css
@@ -1,0 +1,3 @@
+.wp-block-amp-amp-story-post-title {
+	margin: 0;
+}
