+++ assets/src/components/index.js
@@ -14,12 +14,9 @@ export { default as FontFamilyPicker } from './font-family-picker';
 export { default as PreviewPicker } from './preview-picker';
 export { default as getFeaturedImageNotice } from './get-featured-image-notice';
 export { default as getPrePublishNotice } from './get-pre-publish-notice';
-export { default as TextSettings } from './text-settings';
-export { default as ColorSettings } from './color-settings';
 export { default as ResizableBox } from './resizable-box';
 export { default as withMetaBlockEdit } from './with-meta-block-edit';
 export { default as withAmpStorySettings } from './with-amp-story-settings';
-export { default as withAnimationControls } from './with-animation-controls';
 export { default as withAttributes } from './with-attributes';
 export { default as withBlockName } from './with-block-name';
 export { default as withHasSelectedInnerBlock } from './with-has-selected-inner-block';
@@ -27,7 +24,6 @@ export { default as withPageNumber } from './with-page-number';
 export { default as withPrePublishNotice } from './with-pre-publish-notice';
 export { default as withEditFeaturedImage } from './with-edit-featured-image';
 export { default as withStoryFeaturedImageNotice } from './with-story-featured-image-notice';
-export { default as withParentBlock } from './with-parent-block';
 export { default as withIsReordering } from './with-is-reordering';
 export { default as withSelectedBlock } from './with-selected-block';
 export { default as withWrapperProps } from './with-wrapper-props';
