+++ assets/src/components/with-wrapper-props.js
@@ -7,13 +7,12 @@ import { compose } from '@wordpress/compose';
  * Internal dependencies
  */
 import { ALLOWED_BLOCKS, ALLOWED_CHILD_BLOCKS } from '../constants';
-import { withAttributes, withBlockName, withHasSelectedInnerBlock, withParentBlock } from './';
+import { withAttributes, withBlockName, withHasSelectedInnerBlock } from './';
 
 const wrapperWithSelect = compose(
 	withAttributes,
 	withBlockName,
 	withHasSelectedInnerBlock,
-	withParentBlock
 );
 
 /**
