+++ package.json
@@ -25,6 +25,7 @@
     "@wordpress/components": "7.2.0",
     "@wordpress/compose": "3.2.0",
     "@wordpress/data": "4.4.0",
+    "@wordpress/date": "3.2.0",
     "@wordpress/dom-ready": "2.2.0",
     "@wordpress/edit-post": "3.3.0",
     "@wordpress/editor": "9.2.0",
